Add editable model and basic view * Separate config view models and controller * Add remote editable button (wip)
