+++ src/conan_app_launcher/conan_wrapper/unified_api.py
@@ -220,6 +220,11 @@ class ConanUnifiedApiInterface():
     def get_editables_package_path(self, conan_ref: ConanRef) -> Path:
         """ Get package path of an editable reference. """
         raise NotImplementedError
+    
+    @abstractmethod
+    def get_editables_output_folder(self, conan_ref: ConanRef) -> str:
+        """ Get output folder of an editable reference. """
+        raise NotImplementedError
 
     @abstractmethod
     def get_editable_references(self) -> List[str]:
