+++ src/conan_app_launcher/ui/views/conan_conf/conan_conf.ui
@@ -32,7 +32,7 @@
       <string notr="true"/>
      </property>
      <property name="currentIndex">
-      <number>1</number>
+      <number>3</number>
      </property>
      <widget class="QWidget" name="info_widget">
       <property name="styleSheet">
@@ -373,6 +373,13 @@
               </widget>
              </item>
              <item>
+              <widget class="QPushButton" name="remotes_edit_button">
+               <property name="text">
+                <string>Edit</string>
+               </property>
+              </widget>
+             </item>
+             <item>
               <spacer name="remotes_btns_spacer">
                <property name="orientation">
                 <enum>Qt::Vertical</enum>
@@ -415,7 +422,7 @@
       </attribute>
       <layout class="QVBoxLayout" name="verticalLayout_11">
        <item>
-        <layout class="QGridLayout" name="gridLayout_6">
+        <layout class="QGridLayout" name="profiles_grid_layout">
          <item row="0" column="0">
           <widget class="QSplitter" name="splitter">
            <property name="sizePolicy">
@@ -533,9 +540,9 @@
       </attribute>
       <layout class="QVBoxLayout" name="verticalLayout_5">
        <item>
-        <layout class="QGridLayout" name="gridLayout_5">
+        <layout class="QGridLayout" name="editables_grid_layout">
          <item row="0" column="0">
-          <widget class="QSplitter" name="splitter_2">
+          <widget class="QSplitter" name="editables_splitter">
            <property name="sizePolicy">
             <sizepolicy hsizetype="MinimumExpanding" vsizetype="Expanding">
              <horstretch>0</horstretch>
@@ -545,12 +552,11 @@
            <property name="orientation">
             <enum>Qt::Vertical</enum>
            </property>
-           <widget class="QTreeView" name="treeView"/>
-           <widget class="QTreeView" name="treeView_2"/>
+           <widget class="QTreeView" name="editables_ref_view"/>
           </widget>
          </item>
          <item row="0" column="1">
-          <widget class="QFrame" name="frame">
+          <widget class="QFrame" name="editables_button_frame">
            <property name="frameShape">
             <enum>QFrame::StyledPanel</enum>
            </property>
@@ -565,35 +571,35 @@
              <number>0</number>
             </property>
             <item>
-             <widget class="QPushButton" name="pushButton">
+             <widget class="QPushButton" name="editables_refresh_button">
+              <property name="text">
+               <string>Refresh List</string>
+              </property>
+             </widget>
+            </item>
+            <item>
+             <widget class="QPushButton" name="editables_add_button">
               <property name="text">
                <string>Add</string>
               </property>
              </widget>
             </item>
             <item>
-             <widget class="QPushButton" name="pushButton_2">
+             <widget class="QPushButton" name="editables_remove_button">
               <property name="text">
                <string>Remove</string>
               </property>
              </widget>
             </item>
             <item>
-             <widget class="QPushButton" name="pushButton_3">
+             <widget class="QPushButton" name="editables_edit_button">
               <property name="text">
-               <string>PushButton</string>
+               <string>Edit</string>
               </property>
              </widget>
             </item>
             <item>
-             <widget class="QPushButton" name="pushButton_4">
-              <property name="text">
-               <string>PushButton</string>
-              </property>
-             </widget>
-            </item>
-            <item>
-             <spacer name="verticalSpacer_2">
+             <spacer name="editables_button_vspacer">
               <property name="orientation">
                <enum>Qt::Vertical</enum>
               </property>
